// src/components/ui/enums/button-variants.enum.ts
/**
 * Enum representing the different variants of a button.
 */
export enum ButtonVariant {
  Default = 'default',
  Link = 'link',
  Destructive = 'destructive',
  Outline = 'outline',
  Secondary = 'secondary',
  Ghost = 'ghost',
  Underline = 'underline',
  Icon = 'icon'
}

export enum ButtonRadius {
  Default = 'default',
  None = 'none',
  Md = 'md',
  Lg = 'lg',
  Xl = 'xl',
  Full = 'full',
}

export enum ButtonHover {
  Default = 'default',
  Dark = 'dark',
  Icon = 'icon',
  Primary = 'primary',
}